autenticacion en las rutas de notes

+++ src/controllers/note.controllers.js
@@ -1,10 +1,19 @@
 const Note = require('../models/note.model');
 
 const getNoteController = async (req, res, next) => {
+  // se definio en el middleware para hacer la autenticacion
+  const userId = req.user.id;
+
   try {
     // metodo para buscar en la base de datos
-    const allNotes = await Note.find();
-    res.status(200).json(allNotes);
+    // const allNotes = await Note.find();
+
+    // de esta forma solo buscamos las notas del usuario junto con sus datos
+    const notes = await Note.find({
+      user: userId,
+    }).populate('user');
+
+    res.status(200).json(notes);
   } catch (error) {
     // res.status(500).json({ error: error.message });
     next(error);
@@ -14,12 +23,16 @@ const getNoteController = async (req, res, next) => {
 const postNoteController = async (req, res, next) => {
   const { title, content, important } = req.body;
 
+  // se definio en el middleware para hacer la autenticacion
+  const userId = req.user.id;
+
   try {
     // intanciando un nuevo objeto de "Note" y guardando despues
     const newNote = new Note({
       title,
       content,
       important,
+      user: userId, // se hace la referencia al usuario dueño de la nota
     });
 
     const saveNote = await newNote.save();
@@ -31,6 +44,8 @@ const postNoteController = async (req, res, next) => {
 };
 
 const putNoteController = async (req, res, next) => {
+  // se definio en el middleware para hacer la autenticacion
+  const userId = req.user.id;
   const { noteId } = req.params;
   const { title, content, important } = req.body;
 
@@ -42,7 +57,12 @@ const putNoteController = async (req, res, next) => {
     };
 
     // metodo para actualizar en la base de datos tambien existe 'findOneAndUpdate()'
-    const noteUpdate = await Note.findByIdAndUpdate(noteId, newNoteInfo, { new: true });
+    // const noteUpdate = await Note.findByIdAndUpdate(noteId, newNoteInfo, { new: true });
+
+    // Para actualizar la nota solo si le pertenece al usuario logeado
+    const noteUpdate = await Note.findOneAndUpdate({ _id: noteId, user: userId }, newNoteInfo, {
+      new: true,
+    });
 
     // en caso de no encontrar una nota
     if (!noteUpdate) return res.status(404).json({ error: 'No se encontro nota para actualizar' });
@@ -55,12 +75,14 @@ const putNoteController = async (req, res, next) => {
 };
 
 const deleteNoteController = async (req, res, next) => {
+  // se definio en el middleware para hacer la autenticacion
+  const userId = req.user.id;
   const { noteId } = req.params;
 
   try {
     // Utiliza el método 'deleteOne()', 'findOneAndDelete()' o 'findByIdAndDelete()' de Mongoose para eliminar el documento de la base de datos.
     // const deletedNote = await Note.findByIdAndDelete(noteId);
-    const deletedNote = await Note.findOneAndDelete({ _id: noteId });
+    const deletedNote = await Note.findOneAndDelete({ _id: noteId, user: userId });
     if (!deletedNote) return res.status(404).json({ error: 'note not found' });
 
     res.status(200).send('deleted note successful');
